+++ Lua/LuaBindingsManagers.cpp
@@ -354,7 +354,8 @@ namespace RTE {
 		.property("RecommendedMOIDCount", &SettingsMan::RecommendedMOIDCount)
 		.property("AIUpdateInterval", &SettingsMan::GetAIUpdateInterval, &SettingsMan::SetAIUpdateInterval)
 		.property("ShowEnemyHUD", &SettingsMan::ShowEnemyHUD)
-		.property("ScrapCompactingHeight", &SettingsMan::GetScrapCompactingHeight, &SettingsMan::SetScrapCompactingHeight);
+		.property("ScrapCompactingHeight", &SettingsMan::GetScrapCompactingHeight, &SettingsMan::SetScrapCompactingHeight)
+		.property("AutomaticGoldDeposit", &SettingsMan::GetAutomaticGoldDeposit);
 	}
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////