+++ Managers/SettingsMan.h
@@ -256,6 +256,12 @@ namespace RTE {
 		/// </summary>
 		/// <param name="newHeight">The new compacting height, in pixels.</param>
 		void SetScrapCompactingHeight(int newHeight) { m_ScrapCompactingHeight = newHeight; }
+
+		/// <summary>
+		/// Gets whether gold gathered by Actors is automatically added into team funds.
+		/// </summary>
+		/// <returns>Whether gold gathered by Actors is automatically added into team funds.</returns>
+		bool GetAutomaticGoldDeposit() const { return m_AutomaticGoldDeposit; }
 #pragma endregion
 
 #pragma region Network Settings
@@ -512,6 +518,7 @@ namespace RTE {
 		bool m_ShowEnemyHUD; //!< Whether the HUD of enemy actors should be visible to the player.
 		bool m_EnableSmartBuyMenuNavigation; //!< Whether swapping to equipment mode and back should change active tabs in the BuyMenu.
 		int m_ScrapCompactingHeight; //!< The maximum height of a column of scrap terrain to collapse, when the bottom pixel is knocked loose.
+		bool m_AutomaticGoldDeposit; //!< Whether gold gathered by Actors is automatically added into team funds. False means that gold needs to be manually transported into orbit via Craft.
 
 		std::string m_PlayerNetworkName; //!< Player name used in network multiplayer matches.
 		std::string m_NetworkServerAddress; //!< LAN server address to connect to.